LayoutTests/imported/w3c/web-platform-tests/xhr/event-error.sub.any-expected.txt

Blocked access to external URL http://nonexistent.localhost:8800/

PASS onerror should be called
FAIL error while reading body should report zeros for loaded and total assert_equals: loaded expected 0 but got 50
